Adding swipe gestures to navigate thru Onboard; also fixed some issues with Float and CGFLoat (now used in swift 2)

+++ Gulps/ViewControllers/DrinkViewController.swift
@@ -103,7 +103,7 @@ public class DrinkViewController: UIViewController, UIAlertViewDelegate, UIViewC
 
     func updateUI() {
         let percentage = EntryHandler.sharedHandler.currentPercentage()
-        percentageLabel.countFromCurrentValueTo(Float(round(percentage)))
+        percentageLabel.countFromCurrentValueTo(CGFloat(round(percentage)))
         let fillTo = CGFloat(percentage / 100.0)
         progressMeter?.fillTo(fillTo > 1 ? 1 : fillTo)
     }
